feat: implement the initial `message` module (#18)
### What?
This PR introduces the `Message` struct, with it's surrounding (minimal)
set of predefined tags.

### Why?
To streamline the creation of the FIX message and allow further
development, as it lies the foundation for the projects enhancement.

### How?
By introducing a `message` module, in it the message implementation with
it's builder, handling the header and body parts, as well as the `field`
module, taking care of the streamlined FIX field creation.

### Notes
N/A

### Checklist
- [x] Tests are added/updated.
- [x] Documentation is added/updated.
- [x] Self-review of code changes completed.

### Related Issues
Closes #16