feat: implement the initial `Message` encoder (#21)
### What?
This PR introduces a FIX `Message` encoder.

### Why?
To allow users the creation of FIX frame bytes from the already
constructed `Message`.

### How?
By introducing an `encoder` module, exposing the `encode` function which
takes a `Header` and a `Body`, encoding all of their fields following
the FIX protocol field ordering rules. This encoder additionally
auto-populates the tags **9 (BodyLength)** and **10 (Checksum)**.

### Notes
N/A

### Checklist
- [x] Tests are added/updated.
- [x] Documentation is added/updated.
- [x] Self-review of code changes completed.

### Related Issues
Closes #13