feat: implement parser for FIX messages (#15)
Implement a (basic) parser for messages in FIX protocol. This is a
foundational part of any FIX protocol engine, as we must be able to
decode incoming messages.

The decoder assumes (for now) specific FIX version when decoding
messages. Header fields are also not handled in any meaningful way. My
idea on how to handle multiple version of FIX protocol, and also custom
definitions, in the future is through usage of a `trait` and passing in
instance of FIX state that can give us enough meaningful information to
parse the message. For example, this could look something like:

```rust
pub trait FixDecoderState {
    fn is_header_field(&self, field: Field) -> bool;
    fn try_new_field(&self, tag: u16, value: Bytes) -> Result<Field, Error>;
    // ...
}

// and then decode becomes:
fn decode<D: FixDecoderState>(state: D, input: &impl AsRef<[u8]>) -> Result<Message, Error> {
    // ...
}
```

The implemented decoding is very strict as of right now. Meaning, we
expect a slice of bytes(-like) value that contains exactly one message,
with no white-space as prefix and/or suffix.

### Checklist
- [x] Tests are added/updated. (not a lot right now, will be more
extensive when we start supporting specific FIX versions).
- [x] Documentation is added/updated.
- [x] Self-review of code changes completed.

Resolves #11